naver_popular/app/requirements.txt
selenium==4.12.0
webdriver_manager==4.0.1
pandas==2.0.3
requests==2.31.0
uvicorn==0.23.2
fastapi==0.103.1